Move ignore_warnings to sympy.utilities.exceptions
It is needed in some places in library code. But if sympy.testing.pytest is
imported in library code, it will be imported too early for it to detect that
it is running inside of pytest, which breaks the pytest integration. It also
is better there anyway since there are legitimate uses of it in library code
(to prevent the same or similar warnings from issuing twice due to recursive
calls).
