+++ sympy/integrals/transforms.py
@@ -38,8 +38,9 @@ from sympy.polys.rootoftools import CRootOf, RootSum
 from sympy.simplify import simplify, hyperexpand
 from sympy.simplify.powsimp import powdenest
 from sympy.solvers.inequalities import _solve_inequality
-from sympy.testing.pytest import ignore_warnings
-from sympy.utilities.exceptions import sympy_deprecation_warning, SymPyDeprecationWarning
+from sympy.utilities.exceptions import (sympy_deprecation_warning,
+                                        SymPyDeprecationWarning,
+                                        ignore_warnings)
 from sympy.utilities.iterables import iterable
 from sympy.utilities.misc import debug
